login forms created, old logic moved to server side
